Update control.py
Corrected the file reference for initializing details from user (to reference control.py). Hard-coded the code directory because we can't reference the code files without it. (We could move the dialog from user_variables.py to control.py so it prompts the user first--not useful if the goal is making an executable.)